Refactored structure to support module imports better
