+++ converse/modeling/models.py
@@ -1,0 +1,4 @@
+predefined_model_sets = {
+    'orconvqa_default_reader': '',
+    'orconvqa_default_retriever': '',
+}
