+++ converse/src/document_store/elasticsearch.py
