+++ converse/src/document_store/memory.py
